backend_fastapi/requirements.txt
fastapi
uvicorn
motor
pydantic
pydantic-settings
python-dotenv
pymongo
nltk
google-generativeai
python-multipart
httpx
